--- tests/test_executor.py
@@ -1,1 +1,0 @@
-from . import nodes # Using test nodes to test executor
