--- tests/test_nodes.py
@@ -1,1 +1,0 @@
-print("base node tests go here")
